requirements.txt
streamlit==1.28.2
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
Pillow==10.0.1
scikit-learn==1.3.2
joblib==1.3.2
reportlab==4.0.7
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.3
tqdm==4.67.1
